# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Products Documentation",
        "description": "Allows to manage products",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/v1",
            "description": "Test API"
        }
    ],
    "paths": {
        "/products": {
            "post": {
                "summary": "Create Products",
                "description": "Creates new product in the system",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Wireless Keyboard",
                                        "description": "Last-Geneartion Keyboard with Wireless Connection",
                                        "price": 699.98
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                },
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/products/findByName": {
            "get": {
                "description": "Searches for a product by name",
                "summary": "Searches for a product by name",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of searched product",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                },
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Get product by ID",
                "summary": "Get product by ID",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of searched product",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                },
                                "type": "object"
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "id": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securitySchemes": {
        "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}